requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filetype==1.2.0
Flask==3.1.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.44
langchain-google-genai==2.0.11
langchain-openai==0.3.8
langchain-postgres==0.0.13
langsmith==0.3.13
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.66.2
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
proto-plus==1.26.1
protobuf==5.29.3
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.6
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
zstandard==0.23.0
